api: create a specific interface for the device tags
This makes it more organized and easier to make changes
or a new addition to the tag system.
