+++ api/store/device_store.go
@@ -22,11 +22,4 @@ type DeviceStore interface {
 	DeviceSetPosition(ctx context.Context, uid models.UID, position models.DevicePosition) error
 	DeviceListByUsage(ctx context.Context, tenantID string) ([]models.Device, error)
 	DeviceChoice(ctx context.Context, tenantID string, chosen []string) error
-	DeviceCreateTag(ctx context.Context, uid models.UID, tag string) error
-	DeviceDeleteTag(ctx context.Context, uid models.UID, tag string) error
-	DeviceRenameTag(ctx context.Context, tenantID string, currentTagName string, newTagName string) error
-	DeviceListTag(ctx context.Context) ([]string, int, error)
-	DeviceUpdateTag(ctx context.Context, uid models.UID, tags []string) error
-	DeviceGetTags(ctx context.Context, tenantID string) ([]string, int, error)
-	DeviceDeleteAllTags(ctx context.Context, tenantID string, tagName string) error
 }
